Changes to HTML formatting tags
The tags currently in the html.py file are not the best choices for HTML tags. The "tt" tag is not supported in HTML5, and should be converted to "code". The "i" tag should be replaced by the "em" tag. The "b" tag should be replaced by the "strong" tag. The "s" tag should be replaced by the "strike" tag. These are better choices than the current tags in the file.